requirements-openbmc.txt
git+https://github.com/3mdeb/robotframework
robotframework-sshlibrary==3.5.1
robotframework-requests==0.9.4
git+https://github.com/3mdeb/pyte@ignore_param_errors
dbbot-sqlalchemy==0.2
gspread==3.1.0
oauth2client
PyOpenSSL
wakeonlan
cryptography
robotframework-httplibrary
redfish
robotframework-scplibrary
tox
paramiko>=2.5.0
click
PyYAML>=5.1
redfishtool
robotframework-selenium2library
webdrivermanager
robotframework-angularjs
robotframework-xvfb
websocket-client
pyotp
